pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "saadt"
version = "0.0.1"
requires-python = ">=3.12"

dependencies = [
    "beautifulsoup4",
    "bibtexparser>=2.0.0b7",
    "colorlog",
    "dnspython",
    "lxml",
    "PyGObject",
    "PyYAML",
    "regex",
    "requests",
    "Unidecode",
    "urllib3",
]

[project.optional-dependencies]
all = [
    "faust-cchardet",
    "GitPython",
    "numpy",
    "scipy"
]
matching = [
    "GitPython",
]
parser = [
    "numpy",
    "scipy"
]

[tool.setuptools.package-data]
sqlfluff = ["py.typed"]

[tool.mypy]
strict = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"

[[tool.mypy.overrides]]
module = "scipy"
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["B", "E4", "E7", "E9", "F", "I", "N", "PIE", "UP"]
